♻️ refactor(mq-lang): split PathAccess/EnvAccess into io/sandboxed submodules
PathAccess and EnvAccess were both defined inline in sandboxed.rs
alongside SandboxedIo itself. Move each into its own file under
io/sandboxed/ (path_access.rs, env_access.rs) so the module mirrors
the type it groups, keeping sandboxed.rs focused on SandboxedIo.
