+++ crates/mq-lang/src/io/sandboxed.rs
@@ -1,140 +1,12 @@
 use super::{Io, IoError, NativeIo};
 use std::borrow::Cow;
-use std::path::{Component, Path, PathBuf};
+use std::path::{Path, PathBuf};
 
-/// Filesystem access granted for one capability (read or write) of
-/// [`SandboxedIo`].
-///
-/// Built via `impl Into<PathAccess>` so existing call sites keep working
-/// unchanged: `bool` maps to fully denied/allowed, and `Option<Vec<PathBuf>>`
-/// maps directly to a CLI flag parsed as "absent = denied, present with no
-/// paths = fully allowed, present with paths = restricted to those paths"
-/// (e.g. `mq-run`'s `--allow-read`/`--allow-write`).
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub enum PathAccess {
-    /// No filesystem access at all (fail-safe default).
-    #[default]
-    Denied,
-    /// Unrestricted access, matching a bare `--allow-read`/`--allow-write`.
-    Allowed,
-    /// Access restricted to the given files/directories (and their
-    /// descendants).
-    AllowedPaths(Vec<PathBuf>),
-}
+mod env_access;
+mod path_access;
 
-impl From<bool> for PathAccess {
-    fn from(allow: bool) -> Self {
-        if allow { PathAccess::Allowed } else { PathAccess::Denied }
-    }
-}
-
-impl From<Vec<PathBuf>> for PathAccess {
-    fn from(paths: Vec<PathBuf>) -> Self {
-        if paths.is_empty() {
-            PathAccess::Allowed
-        } else {
-            PathAccess::AllowedPaths(paths)
-        }
-    }
-}
-
-impl From<Option<Vec<PathBuf>>> for PathAccess {
-    fn from(paths: Option<Vec<PathBuf>>) -> Self {
-        match paths {
-            None => PathAccess::Denied,
-            Some(paths) => paths.into(),
-        }
-    }
-}
-
-impl PathAccess {
-    fn permits(&self, path: &Path) -> bool {
-        match self {
-            PathAccess::Denied => false,
-            PathAccess::Allowed => true,
-            PathAccess::AllowedPaths(allowed) => {
-                let target = normalize(path);
-                allowed.iter().any(|root| target.starts_with(normalize(root)))
-            }
-        }
-    }
-
-    fn is_denied(&self) -> bool {
-        matches!(self, PathAccess::Denied)
-    }
-}
-
-/// Environment-variable access granted to [`SandboxedIo`]; same shape as [`PathAccess`]
-/// but keyed by variable name instead of filesystem path (e.g. `mq-run`'s `--allow-env`).
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub enum EnvAccess {
-    /// No environment variable access at all (fail-safe default).
-    #[default]
-    Denied,
-    /// Unrestricted access, matching a bare `--allow-env`.
-    Allowed,
-    /// Access restricted to the given variable names.
-    AllowedNames(Vec<String>),
-}
-
-impl From<bool> for EnvAccess {
-    fn from(allow: bool) -> Self {
-        if allow { EnvAccess::Allowed } else { EnvAccess::Denied }
-    }
-}
-
-impl From<Vec<String>> for EnvAccess {
-    fn from(names: Vec<String>) -> Self {
-        if names.is_empty() {
-            EnvAccess::Allowed
-        } else {
-            EnvAccess::AllowedNames(names)
-        }
-    }
-}
-
-impl From<Option<Vec<String>>> for EnvAccess {
-    fn from(names: Option<Vec<String>>) -> Self {
-        match names {
-            None => EnvAccess::Denied,
-            Some(names) => names.into(),
-        }
-    }
-}
-
-impl EnvAccess {
-    fn permits(&self, name: &str) -> bool {
-        match self {
-            EnvAccess::Denied => false,
-            EnvAccess::Allowed => true,
-            EnvAccess::AllowedNames(allowed) => allowed.iter().any(|allowed_name| allowed_name == name),
-        }
-    }
-
-    fn is_denied(&self) -> bool {
-        matches!(self, EnvAccess::Denied)
-    }
-}
-
-fn normalize(path: &Path) -> PathBuf {
-    let absolute = if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        std::env::current_dir().unwrap_or_default().join(path)
-    };
-
-    let mut result = PathBuf::new();
-    for component in absolute.components() {
-        match component {
-            Component::ParentDir => {
-                result.pop();
-            }
-            Component::CurDir => {}
-            other => result.push(other.as_os_str()),
-        }
-    }
-    result
-}
+pub use env_access::EnvAccess;
+pub use path_access::PathAccess;
 
 /// Wraps an inner [`Io`] (typically [`NativeIo`]) with instance-owned
 /// read/write/net/run/env permission grants, enforced by every operation. All
